moved Andthen to arrow-core-data
