+++ modules/core/arrow-core-data/src/test/kotlin/arrow/core/AndThenTest.kt
@@ -1,28 +1,20 @@
-package arrow.data
+package arrow.core
 
 import arrow.Kind
-import arrow.core.andThen
-import arrow.core.compose
 import arrow.core.extensions.monoid
-import arrow.data.extensions.andthen.category.category
-import arrow.data.extensions.andthen.contravariant.contravariant
-import arrow.data.extensions.andthen.monad.monad
-import arrow.data.extensions.andthen.monoid.monoid
-import arrow.data.extensions.andthen.profunctor.profunctor
 import arrow.data.extensions.list.foldable.foldLeft
 import arrow.test.UnitSpec
 import arrow.test.generators.functionAToB
 import arrow.typeclasses.Conested
 import arrow.typeclasses.Eq
-import arrow.typeclasses.conest
 import arrow.typeclasses.counnest
-import arrow.core.identity
+import arrow.data.AndThenOf
+import arrow.data.ForAndThen
 import arrow.test.laws.CategoryLaws
 import arrow.test.laws.ContravariantLaws
 import arrow.test.laws.MonadLaws
 import arrow.test.laws.MonoidLaws
 import arrow.test.laws.ProfunctorLaws
-import io.kotlintest.shouldBe
 import io.kotlintest.properties.Gen
 import io.kotlintest.properties.forAll
 import io.kotlintest.runner.junit4.KotlinTestRunner
